CustomerController and its service are created, users can query all orders of the customer with pagination
